Move the fakeauth out of the Cortex module so it can be reused in other projects.
