+++ pkg/util/fakeauth/fake_auth.go
@@ -1,0 +1,70 @@
+// Package fakeauth provides middlewares thats injects a fake userID, so the rest of the code
+// can continue to be multitenant.
+package fakeauth
+
+import (
+	"context"
+	"net/http"
+
+	"github.com/weaveworks/common/middleware"
+	"github.com/weaveworks/common/server"
+	"github.com/weaveworks/common/user"
+	"google.golang.org/grpc"
+)
+
+// SetupAuthMiddleware for the given server config.
+func SetupAuthMiddleware(config *server.Config, enabled bool, noGRPCAuthOn []string) middleware.Interface {
+	if enabled {
+		config.GRPCMiddleware = append(config.GRPCMiddleware,
+			middleware.ServerUserHeaderInterceptor,
+		)
+		config.GRPCStreamMiddleware = append(config.GRPCStreamMiddleware,
+			func(srv interface{}, ss grpc.ServerStream, info *grpc.StreamServerInfo, handler grpc.StreamHandler) error {
+				for _, path := range noGRPCAuthOn {
+					if info.FullMethod == path {
+						return handler(srv, ss)
+					}
+				}
+				return middleware.StreamServerUserHeaderInterceptor(srv, ss, info, handler)
+			},
+		)
+		return middleware.AuthenticateUser
+	} else {
+		config.GRPCMiddleware = append(config.GRPCMiddleware,
+			fakeGRPCAuthUniaryMiddleware,
+		)
+		config.GRPCStreamMiddleware = append(config.GRPCStreamMiddleware,
+			fakeGRPCAuthStreamMiddleware,
+		)
+		return fakeHTTPAuthMiddleware
+	}
+}
+
+var fakeHTTPAuthMiddleware = middleware.Func(func(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		ctx := user.InjectOrgID(r.Context(), "fake")
+		next.ServeHTTP(w, r.WithContext(ctx))
+	})
+})
+
+var fakeGRPCAuthUniaryMiddleware = func(ctx context.Context, req interface{}, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (interface{}, error) {
+	ctx = user.InjectOrgID(ctx, "fake")
+	return handler(ctx, req)
+}
+
+var fakeGRPCAuthStreamMiddleware = func(srv interface{}, ss grpc.ServerStream, _ *grpc.StreamServerInfo, handler grpc.StreamHandler) error {
+	ctx := user.InjectOrgID(ss.Context(), "fake")
+	return handler(srv, serverStream{
+		ctx:          ctx,
+		ServerStream: ss,
+	})
+}
+
+type serverStream struct {
+	ctx context.Context
+	grpc.ServerStream
+}
+
+func (ss serverStream) Context() context.Context {
+	return ss.ctx
+}
